[BEAM-918] Move StorageLevelPTransform in translation package, use Collections.singletonList() and fix checkstyle
